+++ pcap4j-core/src/main/java/org/pcap4j/core/PacketListener.java
@@ -1,13 +1,11 @@
 /*_##########################################################################
   _##
-  _##  Copyright (C) 2012-2016  Pcap4J.org
+  _##  Copyright (C) 2012-2017  Pcap4J.org
   _##
   _##########################################################################
 */
 
 package org.pcap4j.core;
-
-import org.pcap4j.packet.Packet;
 
 /**
  * @author Kaito Yamada
@@ -20,6 +18,6 @@ public interface PacketListener {
    *
    * @param packet packet
    */
-  public void gotPacket(Packet packet);
+  public void gotPacket(PcapPacket packet);
 
 }